refactor: modularize core runtime primitives
